Move view to evaluation app
