+++ app/grandchallenge/algorithms/templates/algorithms/algorithm_form.html
@@ -3,7 +3,7 @@
 {% load url %}
 
 {% block title %}
-        {{ object|yesno:"Update,Create"}} Algorithm {% if request.resolver_match.view_name == 'algorithms:update' %} Settings {% elif request.resolver_match.view_name == 'algorithms:description-update' %} Description {% elif request.resolver_match.view_name == 'algorithms:create-for-phase' %} for Challenge {% endif %}- {{ block.super }}
+        {{ object|yesno:"Update,Create"}} Algorithm {% if request.resolver_match.view_name == 'algorithms:update' %} Settings {% elif request.resolver_match.view_name == 'algorithms:description-update' %} Description {% elif request.resolver_match.view_name == 'evaluation:phase-algorithm-create' %} for Challenge {% endif %}- {{ block.super }}
 {% endblock %}
 
 {% block breadcrumbs %}
@@ -20,7 +20,7 @@
 
 {% block content %}
 
-    <h2>{{ object|yesno:"Update,Create" }} Algorithm{% if request.resolver_match.view_name == 'algorithms:update' %} Settings {% elif request.resolver_match.view_name == 'algorithms:description-update' %} Description {% elif request.resolver_match.view_name == 'algorithms:create-for-phase' %} for Challenge {% endif %}</h2>
+    <h2>{{ object|yesno:"Update,Create" }} Algorithm{% if request.resolver_match.view_name == 'algorithms:update' %} Settings {% elif request.resolver_match.view_name == 'algorithms:description-update' %} Description {% elif request.resolver_match.view_name == 'evaluation:phase-algorithm-create' %} for Challenge {% endif %}</h2>
 
     {% crispy form %}
